[NU-2189] KafkaSourceFactory removal
