--- engine/flink/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/consumerrecord/FixedValueDeserializationSchemaFactory.scala
@@ -1,23 +1,0 @@
-package pl.touk.nussknacker.engine.kafka.consumerrecord
-
-import org.apache.flink.api.common.serialization.DeserializationSchema
-import org.apache.kafka.common.serialization.{Deserializer, StringDeserializer}
-import pl.touk.nussknacker.engine.kafka.KafkaConfig
-
-/**
-  * Wrapper for value-only Flink's DeserializationSchema. For kafka event key data it uses simple "Array[Byte] to String" deserialization.
-  * Used with simple, value-only, sources where event key is empty or ignored.
-  * It is similar to ConsumerRecordDeserializationSchemaFactory.fixedValueDeserialization but doesn't lost Flink's TypeInformation
-  *
-  * @param valueSchema - value deserialization schema (e.g. EspDeserializationSchema)
-  * @tparam V - type of value of deserialized ConsumerRecord
-  */
-class FixedValueDeserializationSchemaFactory[V](valueSchema: DeserializationSchema[V])
-    extends ConsumerRecordDeserializationSchemaFactory[String, V] {
-  override protected def createKeyDeserializer(kafkaConfig: KafkaConfig): Deserializer[String] = new StringDeserializer
-
-  override protected def createValueDeserializer(kafkaConfig: KafkaConfig): Deserializer[V] = new Deserializer[V] {
-    override def deserialize(topic: String, data: Array[Byte]): V = valueSchema.deserialize(data)
-  }
-
-}
