+++ utils/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/source/KafkaTestParametersInfo.scala
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.engine.kafka.source
+
+import pl.touk.nussknacker.engine.api.definition.Parameter
+import pl.touk.nussknacker.engine.api.test.TestRecord
+
+case class KafkaTestParametersInfo(parametersDefinition: List[Parameter], createTestRecord: Any => TestRecord)
