+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/SchemaBasedSerdeProvider.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.schemedkafka.schemaregistry
 
-import pl.touk.nussknacker.engine.kafka.RecordFormatterFactory
 import pl.touk.nussknacker.engine.schemedkafka.schema.SchemaValidator
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.UniversalToJsonFormatterFactory
 import pl.touk.nussknacker.engine.schemedkafka.serialization.{
   KafkaSchemaBasedDeserializationSchemaFactory,
   KafkaSchemaBasedSerializationSchemaFactory
@@ -10,6 +10,6 @@ import pl.touk.nussknacker.engine.schemedkafka.serialization.{
 case class SchemaBasedSerdeProvider(
     serializationSchemaFactory: KafkaSchemaBasedSerializationSchemaFactory,
     deserializationSchemaFactory: KafkaSchemaBasedDeserializationSchemaFactory,
-    recordFormatterFactory: RecordFormatterFactory,
+    recordFormatterFactory: UniversalToJsonFormatterFactory,
     schemaValidator: SchemaValidator
 )
